requirements.txt
torch
flask
pillow
realesrgan
basicsr
numpy
gdown
flask-cors
